ml-ranking-system/requirements.txt
# Python Dependencies for ML Ranking System

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# ML & Data Science
xgboost==2.0.3
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# NLP & Embeddings
sentence-transformers==2.7.0
torch==2.1.1
transformers==4.35.2

# API & Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1

# Data Generation
faker==21.0.0

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0

# Monitoring
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7
